feat(dg): create dataStore
